Исправления формата логирования и улучшения в bot.py
Внесены изменения в файл `bot.py`:
- Исправлен формат логирования: заменено `levellevel` на `levelname`.
- Добавлена логика извлечения идентификатора топика из URL в функции `process_get_url_step`. Файл теперь сохраняется с именем, основанным на идентификаторе, или отправляется сообщение об ошибке, если идентификатор не найден.
- Добавлен код для запуска бота и логирования его статуса в функции `cancel_search`.

+++ bot.py
@@ -15,7 +15,7 @@ def setup_logging():
     log_level = os.getenv('LOG_LEVEL', 'INFO').upper()
     log_file = os.getenv('LOG_FILE')
     use_console = os.getenv('USE_CONSOLE', 'false').lower() == 'true'
-    log_format = os.getenv('LOG_FORMAT', '%(asctime)s - %(name)s - %(levellevel)s - %(message)s')
+    log_format = os.getenv('LOG_FORMAT', '%(asctime)s - %(name)s - %(levelname)s - %(message)s')
     unauthorized_log_file = os.getenv('UNAUTHORIZED_LOG_FILE')
     enable_unauthorized_logging = os.getenv('ENABLE_UNAUTHORIZED_LOGGING', 'true').lower() == 'true'
 
@@ -153,14 +153,18 @@ def process_get_url_step(message):
     status_message = bot.send_message(message.chat.id, "⏳ Загружаю торрент-файл... Пожалуйста, подождите.")
     torrent_content = rutracker_api.download_torrent_by_url(url)
     if torrent_content:
-        # Получение оригинального названия файла
-        original_filename = url.split('/')[-1]
-        file_path = os.path.join(SAVE_DIRECTORY, original_filename)
-        with open(file_path, 'wb') as f:
-            f.write(torrent_content)
-        os.chmod(file_path, 0o755)
-        bot.delete_message(chat_id=message.chat.id, message_id=status_message.message_id)
-        bot.send_document(message.chat.id, torrent_content, visible_file_name=original_filename, caption="✅ Вот ваш торрент-файл!")
+        # Извлечение id топика из URL
+        topic_id_match = re.search(r't=(\d+)', url)
+        if topic_id_match:
+            topic_id = topic_id_match.group(1)
+            file_path = os.path.join(SAVE_DIRECTORY, f"{topic_id}.torrent")
+            with open(file_path, 'wb') as f:
+                f.write(torrent_content)
+            os.chmod(file_path, 0o755)
+            bot.delete_message(chat_id=message.chat.id, message_id=status_message.message_id)
+            bot.send_document(message.chat.id, torrent_content, visible_file_name=f"{topic_id}.torrent", caption="✅ Вот ваш торрент-файл!")
+        else:
+            bot.edit_message_text(chat_id=message.chat.id, message_id=status_message.message_id, text="❌ Не удалось извлечь id топика из ссылки.")
     else:
         bot.edit_message_text(chat_id=message.chat.id, message_id=status_message.message_id, text="❌ Ошибка при загрузке торрент-файла. Пожалуйста, попробуйте ещё раз позже.")
 
@@ -281,4 +285,5 @@ def cancel_search(call):
 if __name__ == "__main__":
     logging.info("Бот запущен")
     bot.polling(none_stop=True)
+
